requirements.txt
asgiref
crispy-bootstrap5
dj-database-url
Django
django-crispy-forms
django-debug-toolbar
django-email
django-emails
django-environ
django-filter
djangorestframework
gunicorn
Markdown
packaging
pillow
psycopg2
psycopg2-binary
sqlparse
typing_extensions
tzdata